feat: add standalone renderer with comprehensive AST conversion and rendering logic
