+++ apps/core/index.ts
@@ -10,6 +10,7 @@ export type { CompletionItem, CompletionContext } from './completions.js';
 
 export type { DocumentSnapshot, EngineMessage } from './protocol.js';
 export { buildStandaloneExportHtml } from './export-standalone-html.js';
+export { getStandaloneRendererJs } from './standalone-renderer.js';
 
 // Render module
 export * from './render/index.js';
